refactor(config): isolate config decorator shim
Move the dependency-light vLLM-compatible config decorator into a dedicated module. Keep AscendConfig focused on configuration definitions and retain import-cycle regression coverage.
